--- signin.js
@@ -1,9 +1,0 @@
-import login from './utils/loginCheck.js';
-import loginButtonByEnter from './utils/loginButtonEvent.js';
-import emailValidate from './utils/emailValidate.js';
-import passwordValidate from './utils/passwordValidate.js';
-
-login;
-loginButtonByEnter;
-emailValidate;
-passwordValidate;
