+++ signin/signin.html
@@ -11,12 +11,12 @@
       <section class="top">
         <div class="logo">
           <a href="/">
-            <img class="logo" src="./source/logo.png" alt="Linkbrary로고" />
+            <img class="logo" src="../source/logo.png" alt="Linkbrary로고" />
           </a>
         </div>
         <p class="question">회원이 아니신가요? <a href="/signup">회원 가입하기</a></p>
       </section>
-      <form class="middle" name="formname" action="./folder">
+      <form class="middle" action="./folder">
         <div class="input">
           <p>이메일</p>
           <input id="email" type="email" placeholder="codeit@codeit.com" />
@@ -33,8 +33,8 @@
         <div class="socialLogin">
           <p>소셜 로그인</p>
           <div class="sns">
-            <img src="./source/chrome.svg" href="" alt="크롬 아이콘" />
-            <img src="./source/kakao.svg" href="" alt="카카오 아이콘" />
+            <img src="../source/chrome.svg" href="" alt="크롬 아이콘" />
+            <img src="../source/kakao.svg" href="" alt="카카오 아이콘" />
           </div>
         </div>
       </section>
